move all API routes to /routes/api.js
