+++ routes/pages.js
@@ -1,6 +1,6 @@
 var express     = require('express'),
     mock_data   = require('../mock_data/task-data'),
-    pages       = express.Router(),
+    router      = express.Router(),
     debug       = require('debug')('route');
 
 
@@ -13,18 +13,18 @@ var selectPage = function (tab) {
     }
 };
 
-pages.get('/', function (req, res, next) {
+router.get('/', function (req, res, next) {
     res.redirect('/explore');
 });
 
-pages.get('/explore', function (req, res) {
+router.get('/explore', function (req, res) {
     res.render('index', {
         title: "Home",
         page: selectPage('explore')
     });
 });
 
-pages.get('/mytasks', function (req, res) {
+router.get('/mytasks', function (req, res) {
     var myTasks = mock_data._mock_myTask.myTaskIdList.map(function(taskId, idx){
         return mock_data._mock_allTasks[taskId]
     });
@@ -34,16 +34,16 @@ pages.get('/mytasks', function (req, res) {
     });
 });
 
-pages.get('/create', function (req, res) {
+router.get('/create', function (req, res) {
     res.render('create', {
         page: selectPage('create')
     });
 });
 
-pages.get('/setting', function (req, res) {
+router.get('/setting', function (req, res) {
     res.render('setting', {
         page: selectPage('setting')
     });
 });
 
-module.exports = pages;
+module.exports = router;
